requirements.txt
aiosqlite==0.19.0
alembic==1.11.3
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.101.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
pydantic==2.2.1
pydantic_core==2.6.1
redis==5.0.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
asyncpg==0.28.0
yookassa==3.0.1
python-jose==3.3.0
passlib==1.7.4
email-validator==2.0.0
bcrypt==4.0.1
python-jose==3.3.0
python-multipart==0.0.20
rsa==4.9